requirements.txt
# IdleHunter - Python Dependencies
# Python 3.12+ recommended

# ---- Django & Web ----
Django>=5.0,<6
django-environ>=0.11.0
gunicorn>=21.0
whitenoise>=6.6.0

# ---- Database ----
psycopg[binary]>=3.1

# ---- Task Queue & Cache ----
celery[redis]>=5.3.0
redis>=5.0.0
django-celery-results>=2.5.0
django-celery-beat>=2.5.0

# ---- Authentication & Security ----
django-auth-ldap>=2.0.0
#django-auth-ldap>=4.0.0
django-mfa2>=2.0.0
django-allauth>=0.57.0

# ---- API & Swagger ----
djangorestframework>=3.14.0
drf-yasg>=1.21.0

# ---- API Integrations ----
pyvmomi>=8.0.0.0.0
requests>=2.31.0
urllib3>=2.0.0

# ---- ML & Anomaly Detection ----
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# ---- Reports & Export ----
weasyprint>=60.0
reportlab>=4.0.0
openpyxl>=3.1.0

# ---- i18n (gettext is stdlib) ----
# Optional: babel for extra i18n utilities
Babel>=2.13.0

# ---- Development & Quality (optional, can move to requirements/dev.txt) ----
# pytest>=7.4.0
# pytest-django>=4.5.0
# black>=23.0.0
# flake8>=6.1.0
# django-extensions>=3.2.0
# django-debug-toolbar>=4.2.0
